requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
loguru==0.7.2

# Telegram
telethon==1.34.0
cryptg==0.4.0

# AI - MUHIM: openai versiyasini pasaytirdim
openai==1.12.0  # 1.30.0 emas, 1.12.0 ishlatamiz
tenacity==8.2.3

# Database
sqlalchemy==2.0.27
aiosqlite==0.19.0

# Utils
httpx==0.25.0  # 0.27.0 emas